Rename file consistent with other providers
