+++ tests/platform_helper/providers/test_config.py
@@ -8,8 +8,8 @@ import yaml
 from dbt_platform_helper.constants import PLATFORM_CONFIG_FILE
 from dbt_platform_helper.domain.config_validator import ConfigValidator
 from dbt_platform_helper.providers.config import ConfigProvider
-from dbt_platform_helper.providers.yaml_file_provider import DuplicateKeysException
-from dbt_platform_helper.providers.yaml_file_provider import YamlFileProvider
+from dbt_platform_helper.providers.yaml_file import DuplicateKeysException
+from dbt_platform_helper.providers.yaml_file import YamlFileProvider
 from tests.platform_helper.conftest import FIXTURES_DIR
